# Use JSON::Coder when available