--- a/activesupport/lib/active_support/json/encoding.rb
+++ b/activesupport/lib/active_support/json/encoding.rb
@@ -97,6 +97,51 @@ module ActiveSupport
           end
       end
 
+      if defined?(::JSON::Coder)
+        class JSONGemCoderEncoder # :nodoc:
+          JSON_NATIVE_TYPES = [Hash, Array, Float, String, Symbol, Integer, NilClass, TrueClass, FalseClass].freeze
+          CODER = ::JSON::Coder.new do |value|
+            json_value = value.as_json
+            # Handle objects returning self from as_json
+            if json_value.equal?(value)
+              next ::JSON::Fragment.new(::JSON.generate(json_value))
+            end
+            # Handle objects not returning JSON-native types from as_json
+            count = 5
+            until JSON_NATIVE_TYPES.include?(json_value.class)
+              raise SystemStackError if count == 0
+              json_value = json_value.as_json
+              count -= 1
+            end
+            json_value
+          end
+
+
+          def initialize(options = nil)
+            @options = options ? options.dup.freeze : {}.freeze
+          end
+
+          # Encode the given object into a JSON string
+          def encode(value)
+            value = value.as_json(@options) unless @options.empty?
+
+            json = CODER.dump(value)
+
+            # Rails does more escaping than the JSON gem natively does (we
+            # escape \u2028 and \u2029 and optionally >, <, & to work around
+            # certain browser problems).
+            if @options.fetch(:escape_html_entities, Encoding.escape_html_entities_in_json)
+              json.gsub!(">", '\u003e')
+              json.gsub!("<", '\u003c')
+              json.gsub!("&", '\u0026')
+            end
+            json.gsub!("\u2028", '\u2028')
+            json.gsub!("\u2029", '\u2029')
+            json
+          end
+        end
+      end
+
       class << self
         # If true, use ISO 8601 format for dates and times. Otherwise, fall back
         # to the Active Support legacy format.
@@ -126,7 +171,12 @@ module ActiveSupport
 
       self.use_standard_json_time_format = true
       self.escape_html_entities_in_json  = true
-      self.json_encoder = JSONGemEncoder
+      self.json_encoder =
+        if defined?(::JSON::Coder)
+          JSONGemCoderEncoder
+        else
+          JSONGemEncoder
+        end
       self.time_precision = 3
     end
   end