@byroot 2025-02-10
I also crafted a benchmark using Active Record objects, the difference isn't as big but there's a nice speedup there too (~10% faster): https://gist.github.com/byroot/5de28fc54a0cb1284b885f4358a90db7

But there too I suspect `gsub!` is the main bottleneck.