@etiennebarrie 2025-02-12
While working on the deprecation to the special handling of as_json returning self, I realized we also want to allow `JSON::Fragment` as a JSON-native type: https://github.com/rails/rails/pull/54503.